refactor(platform): extract posthog relay helpers
